refactor(PlayersCreator) : create 메서드 수정
