+++ src/test/java/domain/gamer/MoneyTest.java
@@ -1,6 +1,5 @@
-package domain;
+package domain.gamer;
 
-import domain.gamer.Money;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
